Fixes in OOP
